Saving the executed pipeline to disk for reuse in Raidionics [skip ci]

+++ raidionicsrads/Pipelines/ClassificationStep.py
@@ -66,7 +66,7 @@ class ClassificationStep(AbstractPipelineStep):
         return self._patient_parameters
 
     def cleanup(self):
-        if os.path.exists(self._working_folder):
+        if self._working_folder is not None and os.path.exists(self._working_folder):
             shutil.rmtree(self._working_folder)
 
     def __perform_classification(self) -> None:
